requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
Werkzeug==2.3.7
